clean of project basis
